+++ ChangeLog.md
@@ -8,12 +8,13 @@
 |`@autorest/modelerfour` | `4.19.1`
 |`azure-core` dep of generated code | `1.16.0`
 |`msrest` dep of generated code | `0.6.21`
-|`azure-mgmt-core` dep of generated code (If generating mgmt plane code) | `1.2.1`
+|`azure-mgmt-core` dep of generated code (If generating mgmt plane code) | `1.3.0`
 
 **New Features**
 
 - We have added a **provisional** `rest` layer to our generated code. We have also added the following **provisional** flags listed [here](https://github.com/Azure/autorest.python/wiki/Generating-Low-Level-Client#generate-a-low-level-client). #875
 - With this new release, we are also dropping support for Python 3.5 + async.  #875
+- For mgmt plan SDK, default policy changes from `BearerTokenCredentialPolicy` to `ARMChallengeAuthenticationPolicy`.
 
 ### 2021-07-13 - 5.8.4